+++ src/libs/seds/SedsModel/components/activities/splinepoint.h
@@ -19,6 +19,8 @@
 
 #pragma once
 
+#include <cstdint>
+
 namespace seds::model {
 
 class SplinePoint
@@ -29,8 +31,8 @@ public:
     SplinePoint &operator=(SplinePoint &&) = default;
 
 public:
-    auto order() const -> int;
-    auto setOrder(int order) -> void;
+    auto order() const -> uint8_t;
+    auto setOrder(uint8_t order) -> void;
 
     auto raw() const -> double;
     auto setRaw(double raw) -> void;
@@ -39,7 +41,7 @@ public:
     auto setCalibrated(double calibrated) -> void;
 
 private:
-    int m_order;
+    uint8_t m_order;
     double m_raw;
     double m_calibrated;
 };
